refactor: 💡 extract setTransform
